models/team: Move `is_gh_org_owner()` fn into controller
This fn is unrelated to the database, so it should not live in the `models` module.
